Rewrite for flare specifc requirements.
Most of the original c++ files are found in the bestEnclosingRect folder.
The purpose of that program is to find the best enclosing rectangle given
a set of rectangles, which must all be fit into the enclosing boundary.

In the tools flare folder a flare specifc python script can be
found. That script parses flares animation definition file and opens
the images belonging to that animation definition, extracts the
rectangles and passes these to the generic rectangle finding
algorithm. (the compiled tool must be accessible via $PATH)
After the optimization has been performed the results are written
back into the image and definition files.